refactor: address PR #449 review feedback
