+++ wallet/src/lib.rs
@@ -41,7 +41,7 @@ pub mod cli;
 pub mod config;
 pub mod helperfunctions;
 pub mod poller;
-pub mod privacy_preserving_tx;
+mod privacy_preserving_tx;
 pub mod program_facades;
 
 pub const HOME_DIR_ENV_VAR: &str = "NSSA_WALLET_HOME_DIR";
